deleted side bar
